requirements.txt
django-classy-tags==0.6.2
django-crispy-forms==1.5.2
django-db-log==2.2.1
django-filer==0.9.12
django-jet==0.1.3
django-dashing==1.3.0
easy-thumbnails==2.2
PIL==1.1.7
Pillow==3.0.0
wheel==0.24.0

pip install django-datetime-widget
